requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
alembic
streamlit
